Tower

+++ Assets/3.Scripts/Contents/Bullet/Bullet.cs
@@ -6,8 +6,7 @@ public class Bullet : MonoBehaviour
 {
     [HideInInspector]
     public float damage;
-    public LayerMask playerLayer;
-    public LayerMask budildLayer;
+
     private void Start()
     {
         Destroy(gameObject,5);
@@ -15,7 +14,7 @@ public class Bullet : MonoBehaviour
 
     private void OnTriggerEnter2D(Collider2D collision)
     {
-        if(collision.gameObject.layer == Mathf.Log(playerLayer.value,2) || collision.gameObject.layer == Mathf.Log(budildLayer.value, 2))
+        if(collision.GetComponent<IGetDamage>() != null)
         {
             Hit(collision);
         }
@@ -23,16 +22,30 @@ public class Bullet : MonoBehaviour
 
     void Hit(Collider2D col)
     {
-        col.GetComponent<Stat>().Hp -= damage; 
-        if(col.GetComponent<Stat>().Hp <= 0)
+        if (col.GetComponent<IGetDamage>() != null)
         {
-            if (col.GetComponent<Item>())
+            if (col.GetComponent<Stat>() != null)
             {
-                col.GetComponent<Item_Buliding>().DeleteBuilding();
+                col.GetComponent<Stat>().Hp -= damage;
+                if (col.GetComponent<Stat>().Hp <= 0)
+                {
+                    if (col.GetComponent<Item>())
+                    {
+                        col.GetComponent<Item_Buliding>().DeleteBuilding();
+                    }
+                    else if (col.GetComponent<TowerController>())
+                    {
+                        Debug.Log("±âÁö ÆÄ±«");
+                    }
+                }
             }
-            else if (col.GetComponent<TowerController>())
+            else if (col.GetComponent<PlayerStat>() != null)
             {
-                Debug.Log("±âÁö ÆÄ±«");
+                col.GetComponent<PlayerStat>().Hp -= damage;
+                if (col.GetComponent<PlayerStat>().Hp <= 0)
+                {
+                    Debug.Log("ÇÃ·¹ÀÌ¾î Á×À½");
+                }
             }
         }
         Destroy(gameObject);
